requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
beautifulsoup4==4.12.2
readability-lxml==0.8.1
anthropic==0.86.0
openai==2.29.0
google-generativeai==0.8.6
networkx==3.2.1
pydantic==2.5.2
python-multipart==0.0.6
python-Levenshtein==0.23.0
numpy==2.4.3
scipy==1.17.1
pytest==8.4.1
pytest-asyncio==1.1.0
